laneige-insight-mvp/requirements.txt
playwright==1.48.0
beautifulsoup4==4.12.3
lxml==5.3.0
httpx==0.27.2
pydantic==2.9.2
python-dotenv==1.0.1
SQLAlchemy==2.0.36
Pillow==10.4.0
imagehash==4.3.1
pandas==2.2.3
streamlit==1.39.0
